requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
google-generativeai
python-dotenv
faiss-cpu
langchain-community
